kube: rename resource mod to request
since it no longer contains a Resource struct after rename
